backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database - Qdrant (1.7.1+ supports Python 3.13)
qdrant-client>=1.7.1

# Database - Postgres with async
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1

# OpenAI Agents SDK
openai-agents>=0.0.3
openai>=1.12.0
tiktoken>=0.5.2

# Utilities
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0

# Logging and monitoring
structlog>=24.1.0

# Authentication
# NOTE: python-jose[cryptography] is required for JWT
# passlib provides secure password hashing abstraction
# bcrypt 4.1.2 is compatible with passlib (5.x breaks __about__ attribute)
bcrypt==4.1.2
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
